requirements.txt
django
djangorestframework
wheel